+++ app/src/features/editor/editor-menu.tsx
@@ -73,10 +73,6 @@ export default function EditorMenu() {
               <FileText size={18} />
               PDF
             </DropdownMenuItem>
-            <DropdownMenuSeparator />
-            <span className="opacity-80 text-sm m-3">
-              {t("editor.menu.imageWarning")}
-            </span>
           </DropdownMenuSubContent>
         </DropdownMenuSub>
         <DropdownMenuItem onSelect={importer.importNotes}>